mcp-server/requirements.txt
mcp>=1.0.0
asyncpg>=0.29.0
redis>=5.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
